Continue refactoring functions into categories. Lint and imports
