+++ crates/embucket-functions/src/semi-structured/object/object_construct.rs
@@ -206,7 +206,7 @@ impl ScalarUDFImpl for ObjectConstructUDF {
 #[allow(clippy::unwrap_used)]
 mod tests {
     use super::*;
-    use crate::variant::array_construct::ArrayConstructUDF;
+    use crate::semi_structured::array::array_construct::ArrayConstructUDF;
     use datafusion::assert_batches_eq;
     use datafusion::prelude::SessionContext;
     use datafusion_expr::ScalarUDF;
