+++ crates/embucket-functions/src/semi-structured/variant/visitors/array_construct.rs
@@ -54,7 +54,7 @@ pub fn visit(stmt: &mut Statement) {
 #[allow(clippy::unwrap_used)]
 mod tests {
     use super::*;
-    use crate::variant::array_construct::ArrayConstructUDF;
+    use crate::semi_structured::array::array_construct::ArrayConstructUDF;
     use datafusion::assert_batches_eq;
     use datafusion::prelude::SessionContext;
     use datafusion::sql::parser::Statement as DFStatement;
